requirements.txt
tensorflow-cpu
opencv-python-headless
pillow
pandas
numpy
streamlit
streamlit-js-eval